% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hLDA.R
\docType{class}
\name{hLDA.model-class}
\alias{hLDA.model-class}
\title{hLDA Class Definition}
\description{
The hLDA class contains information associated with a hierarchical LDA model.
The available accessor methods are,
}
\details{
* \code{gammas}: Extract the sample composition.
* \code{betas}: Extract the topic composition.
* \code{topics}: Extract the topic information.
* \code{hierarchy}: Extract the hierarchical structure.
* \code{par}: Returns the parameters used for model fitting.
}
\seealso{
hLDA
}
