% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hLDA.R
\name{hLDA}
\alias{hLDA}
\title{Fits a hLDA Model to Count Data}
\usage{
hLDA(
  x,
  depth,
  gamma = 0.1,
  alpha = 0.1,
  eta = 0.1,
  print = c("nothing", "summary", "everything")
)
}
\arguments{
\item{x}{a matrix of counts.
Features are on columns and samples are on rows.}

\item{depth}{(required, \code{integer}) the desired depth of the
hierarchical structure.}

\item{gamma}{(optional, \code{real} in \code{]0,1[}, default = \code{0.1})
the concentration parameter for introducing new topics.}

\item{alpha}{(optional, \code{real} in  \code{]0,1[}, default = \code{0.1})
hyperparameter of Dirichlet distribution for document-depth level.}

\item{eta}{(optional, \code{real} in \code{]0,1[}, default = \code{0.1})
hyperparameter of Dirichlet distribution for topic-word.}

\item{print}{(optional, \code{character}, default = \code{"nothing"}) What
should be printed to the console?}
}
\value{
a \code{hLDA.model} object.
}
\description{
Fits a hLDA Model to Count Data
}
\examples{
x <- matrix(sample(0:100, 1000, replace = TRUE), 20, 50)
m <- hLDA(x, depth = 3)
plot_hLDA(m)
}
