% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hlda.R
\name{plot_hLDA}
\alias{plot_hLDA}
\title{Visualizes the hLDA Model}
\usage{
plot_hLDA(
  hLDA_model,
  reorder_feature = TRUE,
  min_p = 0.01,
  title = "",
  color_by = "path"
)
}
\arguments{
\item{hLDA_model}{(required, \code{hLDA.model}) object returned by \code{hLDA}.}

\item{reorder_feature}{(optional, \code{logical}, default = \code{TRUE})
whether features should be re-ordered to group features belonging to the same
topics in the visualization of the topic composition.}

\item{min_p}{(optional, \code{real} in \code{]0,1[}, default = \code{0.01})
the proportion a feature must have in any topic to be included in the topic
composition visualization.}

\item{title}{(optional, \code{character}, default = \code{""})}

\item{color_by}{(optional, \code{character}, default = \code{"path"}) specifies the
variable that should be used for coloring the visualization.}
}
\description{
Specifically, \code{plot_hLDA} shows the hierarchy
(which can be visualized separately with \code{plot_hLDA_hierarchy}),
the topic composition (see \code{plot_hLDA_betas}), and the sample
composition (see \code{plot_hLDA_gammas} or \code{plot_hLDA_gammas_by_path}).
}
\examples{
x <- matrix(sample(0:100, 1000, replace = TRUE), 20, 50)
m <- hLDA(x, depth = 3)
plot_hLDA(m)
}
\seealso{
plot_hLDA_hierarchy, plot_hLDA_betas, plot_hLDA_gammas
}
