% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hlda_betas.R
\name{plot_hLDA_betas}
\alias{plot_hLDA_betas}
\title{Visualizes Topic Composition of a Fitted hLDA Model}
\usage{
plot_hLDA_betas(
  hLDA_model,
  reorder_feature = TRUE,
  min_p = 0.01,
  color_by = "path",
  print_legend = FALSE
)
}
\arguments{
\item{hLDA_model}{(required, \code{hLDA.model}) object returned by \code{hLDA}.}

\item{reorder_feature}{(optional, \code{logical}, default = \code{TRUE})
whether features should be re-ordered to group features belonging to the same
topics in the visualization of the topic composition.}

\item{min_p}{(optional, \code{real} in \code{]0,1[}, default = \code{0.01})
the proportion a feature must have in any topic to be included in the topic
composition visualization.}

\item{color_by}{(optional, \code{character}, default = "path") specifies the
variable that should be used for coloring the visualization.}

\item{print_legend}{(optional, \code{logical}, default = \code{FALSE})
whether a legend should be printed to the console.}
}
\value{
A \code{ggplot2} object describing the sample composition.
}
\description{
Visualizes Topic Composition of a Fitted hLDA Model
}
\examples{
x <- matrix(sample(0:100, 1000, replace = TRUE), 20, 50)
m <- hLDA(x, depth = 3)
plot_hLDA_betas(m)
}
\seealso{
plot_hLDA
}
