% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hlda_gammas.R
\name{plot_hLDA_gammas}
\alias{plot_hLDA_gammas}
\title{Visualizes Sample (Document) Composition of a Fitted hLDA Model}
\usage{
plot_hLDA_gammas(hLDA_model, color_by = "path", print_legend = FALSE)
}
\arguments{
\item{hLDA_model}{(required, \code{hLDA.model}) object returned by \code{hLDA}.}

\item{color_by}{(optional, \code{character}, default = "path") specifies the
variable that should be used for coloring the visualization.}

\item{print_legend}{(optional, \code{logical}, default = \code{FALSE})
whether a legend should be printed to the console.}
}
\value{
A \code{ggplot2} object describing the sample composition.
}
\description{
Visualizes Sample (Document) Composition of a Fitted hLDA Model
}
\examples{
x <- matrix(sample(0:100, 1000, replace = TRUE), 20, 50)
m <- hLDA(x, depth = 3)
plot_hLDA_gammas(m)
}
\seealso{
plot_hLDA
}
