% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hlda_hierarchy.R
\name{plot_hLDA_hierarchy}
\alias{plot_hLDA_hierarchy}
\title{Visualizes the Topic Hierarchy of a Fitted hLDA Model}
\usage{
plot_hLDA_hierarchy(hLDA_model, direction = "h", color_by = "path")
}
\arguments{
\item{hLDA_model}{(required, \code{hLDA.model}) object returned by \code{hLDA}.}

\item{direction}{(optional, \code{character}, default = \code{"h"})
whether the hierarchy should be displayed vertically (\code{"v"}) or horizontally (\code{"h"}).}

\item{color_by}{(optional, \code{character}, default = \code{"path"}) specifies the
variable that should be used for coloring the visualization.}
}
\value{
A \code{ggplot2} object describing the sample composition.
}
\description{
Visualizes the Topic Hierarchy of a Fitted hLDA Model
}
\examples{
x <- matrix(sample(0:100, 1000, replace = TRUE), 20, 50)
m <- hLDA(x, depth = 3)
plot_hLDA_hierarchy(m)
}
\seealso{
plot_hLDA
}
