% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_count_matrix_to_corpus.R
\name{transform_count_matrix_to_corpus}
\alias{transform_count_matrix_to_corpus}
\title{Transform a count matrix into a corpus data.frame}
\usage{
transform_count_matrix_to_corpus(data)
}
\description{
Transform a count matrix into a corpus data.frame
}
